requirements.txt
openai
rich
prompt_toolkit
ddgs
python-dotenv
pytest
